==> analysis/01_ingest.qmd <==
---
title: Inladen en voorbereiden
---

{{< include 00_setup.qmd >}}

Voordat we met de analyse beginnen, lade we eerst de data in. De functies hiervoor zijn te vinden in R/ingest.R.

Naast het inladen van de data, wordt ook data geprepareerd. Soms is dit een afgeleide variabele maken, maar het kan ook zijn dat de data wordt omgezet naar een ander detailniveau (bijvoorbeeld van een tabel met per regel een status, naar een tabel met per regel een verbintenis en kolommen met info per status). Deze functies zijn de vinden in R/prepare.R.

## Verbintenissen (enrollments)

We laden eerst alle data in die al op verbintenis-niveau is. Een aantal tabellen zijn niet op verbintenis-niveau en bevatten ook geen verbintenis ID. Denk hierbij aan BPV statussen en vooropleiding kenmerken van studenten.

Deze tabellen transformeren we naar student-jaar om het zo te kunnen koppelen aan verbintenis. Tenslotte zijn er nog een aantal mapping tables met extra informatie over een specifieke aspecten van een verbintenis.

Alle volgende ingests zijn complete exports. Als op een later moment het onderliggende SQL script opnieuw wordt gerund (zie Teams map) dan is het automatisch up-to-date.  

```{r}

enrollments_basics <- ingest_enrollments_basics() |>
    #'*INFO* prepare derived variables regarding 'studiesucces' after combining data with programmes
    combine_enrollment_numbers_for_order() |>
    parse_enrollment_dates()
enrollments_level <- ingest_enrollments_level() |>
    parse_enrollment_level()
enrollments_flex <- ingest_enrollments_flex() |>
    create_flex_boolean()
enrollments_special_needs <- ingest_enrollments_special_needs() |>
    summarise_special_needs() |>
    create_special_needs_boolean()
            
enrollments_study_advices <- ingest_study_advices() |>
    transform_advices_to_enrollments()
enrollments_application <- ingest_enrollments_application() |>
    calculate_application_duration()

print("Data op verbintenis niveau succesvol ingeladen en geprepareerd.")

```

```{r}

# BPV statusses transformed
bpv_registrations <- ingest_bpv_registrations()
bpv_statusses <- ingest_bpv_statusses()

enrollments_bpv <- bpv_statusses |>
    left_join(bpv_registrations,
              by = join_by(BPV_ID),
              relationship = "many-to-one") |>
    convert_bpv_status_types() |>
    transform_bpv_statusses_to_enrollments() 

print("BPV data succesvol ingeladen en getransformeerd naar verbintenis.")

```

```{r}
#' *INFO* An enrollment can be longer or shorter than one year
student_demographics_yearly <- ingest_students_demographics() |>
    mutate(DEELNEMER_postcode4 = suppressWarnings(as.numeric(DEELNEMER_postcode4))) |>
    transform_students_to_student_year()
student_prior_education_yearly <- ingest_students_prior_education() |>
    transform_prior_education_to_student_year() |>
    transform_prior_education_vo_and_highest_degree()

print("Demografie en vooropleiding succesvol ingeladen en getransformeerd naar verbintenis en student-schooljaar.")
```

```{r}
# Load postal code related data
# APCG (Armoede Probleem Cumulatie Gebieden) indicates areas that need special attention
# SES data contains socioeconomic status scores from CBS per postal code
student_postcode_apcg <- ingest_student_postcode_apcg()
student_postcode_ses <- ingest_student_postcode_ses() |>
        mutate(
            # Convert SES score and spread to numeric values
            DEELNEMER_postcode4_ses_score = suppressWarnings(as.numeric(DEELNEMER_postcode4_ses_score)),
            DEELNEMER_postcode4_ses_spreiding = suppressWarnings(as.numeric(DEELNEMER_postcode4_ses_spreiding))
        ) |> save_prepared_and_return()

print("Postcode-gerelateerde data (APCG en SES informatie) succesvol ingeladen.")

```

### Aanvullende dimensies

We laden ook een aantal mapping tables in. Deze tabellen bevatten extra informatie over een specifiek aspect van een verbintenis. Denk hierbij aan redenen voor uitschrijving, cohorten, teams en locaties.

Teams en locaties zijn dus in de eerste plaats verbonden aan verbintenissen, niet aan opleidingen. Dit komt omdat een instellingen meerdere opleidingen op verschillende locaties kan hebben.

```{r}

reasons_for_leaving <- ingest_reasons_for_leaving() 
cohorts <- ingest_cohorts() |>
    add_cohort_start_year() |>
    add_cohort_start_date()

teams <- ingest_teams()
locations <- ingest_locations()

print("Aanvullende dimensies succesvol ingeladen en geprepareerd.")

```

### Opleidingen (programmes)

```{r}

programmes_exam_plan <- ingest_exam_plans() |> 
    summarise_plan_dates_to_programmes()
programmes_basics <- ingest_programmes_basics()
programme_bc_codes <- ingest_programme_bc_codes()

print("Programma (incl. examenplan) data succesvol ingeladen en geprepareerd.")

```

## Medewerkers (employees)

We laden eerst informatie uit medewerkers contracten in. Deze data is per jaar aangeleverd met peildatum 1 oktober. Naast de contracten, zijn er ook componenten met FTE aanpassingen.

```{r}

employees_contract <- bind_rows(
    ingest_employees_contract_basics_2023(),
    ingest_employees_contract_basics_2022(),
    ingest_employees_contract_basics_2021(),
    ingest_employees_contract_basics_2020(),
    ingest_employees_contract_basics_2019()
) |> 
    add_school_year() |>
    save_prepared_and_return(filename = "employees_contract")

employees_contract_extra_fte <- bind_rows(
    ingest_job_components_extra_fte_2023(),
    ingest_job_components_extra_fte_2022(),
    ingest_job_components_extra_fte_2021(),
    ingest_job_components_extra_fte_2020(),
    ingest_job_components_extra_fte_2019()
) |>
    summarise_components_to_employees()

print("Medewerker FTE data op peildatum succesvol ingeladen en geprepareerd.")


```

Tenslotte is er informatie over het verzuim van medewerkers. Deze zetten we om naar jaarlijkse verzuimpercentages. Hierbij ontbreken gegeven van medewerkers die geen enkele verzuim melding in een jaar hebben. Dit vullen op na het koppelen.

```{r}

# Within map to keep the transformations separate by years, this makes it less memory intensive
employees_absence_yearly <- ingest_employee_absences() |>
    # Due to joins before the export there are duplicates
    distinct() |>
    # TODO rename function after fixing dates at ingest
    convert_absence_types() |>
    summarise_employee_absence_to_years()

print("Medewerker verzuim data succesvol ingeladen en getransformeerd naar schooljaar.")

```

```{r}

employees_job_type <-  bind_rows(
    ingest_employees_job_type_2023(),
    ingest_employees_job_type_2022(),
    ingest_employees_job_type_2021(),
    ingest_employees_job_type_2020(),
    ingest_employees_job_type_2019()
) |>
    # TODO some double values due to joins before export
    # This unfortunately also removes a few teachers who have double contracts
    distinct(
        MEDEWERKER_ID, 
        TEAM_kostenplaats_code, 
        MEDEWERKER_functie_peildatum, 
        .keep_all = TRUE
  ) |>
    save_prepared_and_return(filename = "employees_job_type")

print("Medewerker contract, fte, functie en verzuim data succesvol ingeladen en geprepareerd.")

```

## Waarnemingen / verzuim (attendance)

M.b.t. aanwezigheid zijn er waarnemingen. Deze worden gekoppeld aan groepsinschrijvingen van studenten. De waarnemeningen data is de grootste dataset dus dit inladen duurt even.

```{r}
#| freeze: true
#| cache: false

attendance_observations <- ingest_attendance_observations() 

enrollment_to_sk_mapping <- ingest_enrollment_to_sk_mapping()
group_placements <- ingest_group_placements()

group_placements_enriched <- group_placements |>
    # Enrollments absences only have SK_inschrijving that started over 2018, therefore this
    # inner join removes half of rows
    inner_join(enrollment_to_sk_mapping,
               by = join_by(SK_inschrijving),
               relationship = "many-to-one")

attendance_observations_enriched <- attendance_observations |>
    # I lose ~2400 enrollments
    # maybe these are students from before 2018, maybe some are filtered in data warehouse 
    inner_join(group_placements_enriched,
                by = join_by(SK_GroepInschrijving == SK_groep_inschrijving),
               relationship = "many-to-one")

## Remove for freeing up memory
rm(attendance_observations)

## Yearly attendance data
enrollment_years_attendance <- attendance_observations_enriched |>
    summarise_observations_to_yearly_attendance()

## Weekly attendance data - first 14 weeks of each school year
enrollment_weeks_attendance <- attendance_observations_enriched |>
    summarise_observations_to_weekly_attendance()

## Transform weekly data to enrollment level with pivoted weeks
enrollments_years_attendance_weekly <- enrollment_weeks_attendance |>
    transform_attendance_weekly_to_enrollments()

print("Waarnemingen data succesvol ingeladen en getransformeerd naar schooljaar en per week.")

```

## Student tevredenheid (satisfaction)

We laden hier de JOB monitor in. In 2019 en 2021 waren er bijna dezelfde vragen, in 2023 waren er veel meer vragen (138 tov 85 en 78). Voor het correlatie-onderzoek kunnen we deze allemaal gebruiken, omdat missende waardes geen probleem zijn.

De Job light heeft ook andere vragen en deze zijn ook anders geformateerd. Voor nu is ervoor gekozen dit niet in te laden.

```{r}
#| warning: false
# CREBO is set as integer which gives warnings when it says "Niet getonond wegens AVG"


students_satisfaction <- bind_rows(
    ingest_students_satisfaction_2019(),
    ingest_students_satisfaction_2021(),
    ingest_students_satisfaction_2023()
) |>
    filter(!is.na(CREBO)) |>
    save_prepared_and_return(filename = "students_satisfaction")

codebook_student_satisfaction_mapping <-  ingest_codebook_students_satisfaction_mapping()
#Based on the context, I'll provide code to transform the students_satisfaction names using the codebook mapping. Here's the code:

rename_mapping_job <- codebook_student_satisfaction_mapping |>
    select(vraag_code_job, omschrijving) |>
    filter(!is.na(omschrijving)) |>
    pull(vraag_code_job, name = omschrijving) |>
    discard(is.na)

students_satisfaction <- students_satisfaction |>
    rename(any_of(rename_mapping_job)) |>
    rename_with(~ paste0("JOB_", .)) |>
    save_prepared_and_return(filename = "students_satisfaction")

print("Student tevredenheid data succesvol ingeladen en geprepareerd.")

crebo_to_organisation_mapping <- ingest_crebo_to_organisation_mapping() |>
    distinct(OPLEIDING_crebo_correctie, .keep_all = TRUE) |>
    save_prepared_and_return()
    
    
```

```{r}

# student_answers_satisfaction <- ingest_student_answers_satisfaction_2020()

students_satisfaction_2020 <- ingest_student_answers_satisfaction_2020() |>
    mutate(attribute = parse_number(attribute)) |>
    pivot_answers_to_students()

rename_mapping_job_light_2021 <- codebook_student_satisfaction_mapping |>
    select(vraag_code_job_light_2021, omschrijving) |>
    mutate(vraag_code_job_light_2021 = paste0("x", vraag_code_job_light_2021)) |>
    filter(!is.na(omschrijving)) |>
    pull(vraag_code_job_light_2021, name = omschrijving) |>
    discard(is.na)

students_satisfaction_2020 <- students_satisfaction_2020 |>
  rename(any_of(rename_mapping_job_light_2021)) |>
    save_prepared_and_return()

```

```{r}
students_satisfaction_2022 <- ingest_student_answers_satisfaction_2022() |>
    mutate(attribute = parse_number(attribute)) |>
    pivot_answers_to_students()

rename_mapping_job_light_2023 <- codebook_student_satisfaction_mapping |>
    select(vraag_code_job_light_2023, omschrijving) |>
    mutate(vraag_code_job_light_2023 = paste0("x", vraag_code_job_light_2023)) |>
    filter(!is.na(omschrijving)) |>
    pull(vraag_code_job_light_2023, name = omschrijving) |>
    discard(is.na)

students_satisfaction_2022 <- students_satisfaction_2022 |>
    rename(any_of(rename_mapping_job_light_2023)) |>
    rename_with(~str_replace(., "x", "x_")) |>
    save_prepared_and_return()

```

## Medewerker tevredenheid (satisfaction)

Laad de medewerker tevredenheid data in. Deze data is per jaar aangeleverd met één antwoord per rij. Laad eerste de codeboeken in en koppel deze aan de antwoorden. Vervolgens worden de antwoorden omgezet naar een tabel met per medewerker één rij en alle gegeven antwoorden in de kolommen.

```{r}

codebook_employee_satisfaction_mapping <-  ingest_codebook_employee_satisfaction_mapping()

codebook_employees_satisfaction_2020 <- ingest_employee_answers_satisfaction_codebook_2020() |>
    left_join(
        codebook_employee_satisfaction_mapping,
        by = c("Question Text" = "vraag"),
        relationship = "many-to-one"
    ) |>
    mutate(omschrijving = coalesce(omschrijving, `Question Text`)) |>
    save_prepared_and_return()

codebook_employees_satisfaction_2022 <- ingest_employee_answers_satisfaction_codebook_2022() |>
        left_join(
        codebook_employee_satisfaction_mapping,
        by = c("Question Text" = "vraag"),
        relationship = "many-to-one"
    ) |>
    mutate(omschrijving = coalesce(omschrijving, `Question Text`)) |>
    save_prepared_and_return()

codebook_employees_satisfaction_2024 <- ingest_employee_answers_satisfaction_codebook_2024() |>
        left_join(
        codebook_employee_satisfaction_mapping,
        by = c("Question Text" = "vraag"),
        relationship = "many-to-one"
    ) |>
    mutate(omschrijving = coalesce(omschrijving, `Question Text`)) |>
    save_prepared_and_return()

employees_satisfaction_2020 <- ingest_employee_answers_satisfaction_2020() |>
    left_join(
        codebook_employees_satisfaction_2020,
        by = c("QuestionId" = "QuestionId", "Score" = "AnswerCode"),
        relationship = "many-to-one"
  ) |>
    pivot_answers_to_employees()

employees_satisfaction_2022 <- ingest_employee_answers_satisfaction_2022() |>
    left_join(
        codebook_employees_satisfaction_2022,
        by = c("QuestionId" = "QuestionId", "Score" = "AnswerCode"),
        relationship = "many-to-one"
    ) |> 
    pivot_answers_to_employees()

employees_satisfaction_2024 <- ingest_employee_answers_satisfaction_2024() |>
    left_join(
        codebook_employees_satisfaction_2024,
        by = c("QuestionId" = "QuestionId", "Score" = "AnswerCode"),
        relationship = "many-to-one"
    ) |> 
    pivot_answers_to_employees()

mapping_employees_satisfaction <- ingest_mapping_employees_satisfaction()

employees_satisfaction <- bind_rows(
    employees_satisfaction_2020, 
    employees_satisfaction_2022, 
    employees_satisfaction_2024
    ) |>
    left_join(
        mapping_employees_satisfaction,
        by = join_by(MWO_characteristic_1, MWO_characteristic_2),
        relationship = "many-to-one"
    ) |> 
    save_prepared_and_return(filename = "employees_satisfaction")


print("Medewerker tevredenheid data succesvol ingeladen en geprepareerd.")


```

## Rendementen

Voor nu is het startersresultaat (retention start). Dit zijn studenten die 1 jaar in de opleiding zitten. Zo kunnen we variabelen die pear jaar verschillen zoals medewerkerstevredenheid en studenttevredenheid hieraan koppelen.

Daarbij is er het jaarresultaat (year). Dit is het aantal vertrekkers van de instelling gedeeld door aantal vertrekkkers met diploma in dat jaar.

En tenslotte het diplomaresultaat (degree). Dit is het aantal vertrekkers gedeeld door aantal vertrekkers die een diploma bij MBOA hebben gehaald). Oftewel bij het diplomaresultaat tellen ook studenten mee die in 5 eerdere jaren een diploma hebben gehaald, doorstuderen bij MBO Amersfoort en dan bij nieuwe opleiding uitvallen zonder diploma.

Zie de MBO Informatie Encylopedie voor precieze definitie en voorbeelden:

-   [Startersresultaat](https://www.informatie-encyclopedie.nl/indicator/4fcb1f2b-2c31-11b2-8056-00505687193f)
-   [Diplomaresultaat](https://www.informatie-encyclopedie.nl/indicator/27693a2c-23ad-4d12-871b-10edecc4e6a4)
-   [Jaarresultaat](https://www.informatie-encyclopedie.nl/indicator/ed395a2e-d2b7-42d1-b726-b1d091ca73f4)

```{r}

teams_results_retention_start <- ingest_teams_results_retention_start() |>
    format_school_year_name() |>
    parse_result_pct()

teams_results_year <- ingest_teams_results_year() |>
    format_school_year_name() |>
    parse_result_pct()

teams_results_degree <- ingest_teams_results_degree() |>
    format_school_year_name() |>
    parse_result_pct()

print("Startersresultaat succesvol ingeladen en geprepareerd.")

```
